requirements.txt
pyodbc>=5.0.0
azure-identity>=1.15.0
pandas>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
